+++ trabahakita.com/app/Http/Controllers/SkillsController.php
@@ -13,51 +13,62 @@ class SkillsController extends Controller
 
     public function store(Request $request)
     {
-        $this->validate(
-            request(),
-            [
-                'skill' => 'required',
+        if (auth()->check()) {
+            $this->validate(
+                request(),
+                [
+                    'skill' => 'required',
+                ]
+            );
 
-            ]
-        );
+            $skill = new Skills;
+            $skill->user_id = auth()->user()->id;
+            $skill->desc = request('skill');
+            $skill->save();
 
-        $skill = new Skills;
-        $skill->user_id = auth()->user()->id;
-        $skill->desc = request('skill');
-        $skill->save();
-
-        $profiles = Profile::find(auth()->user()->id);
-        $education = Education::all();
-        $skills = Skills::all();
-        return redirect()->to('/employee/education')->with(['profile' => $profiles, 'education' => $education, 'skills' => $skills]);
+            $profiles = Profile::find(auth()->user()->id);
+            $education = Education::all();
+            $skills = Skills::all();
+            return redirect()->to('/employee/education')->with(['profile' => $profiles, 'education' => $education, 'skills' => $skills]);
+        } else {
+            return redirect()->to('/login');
+        }
     }
 
     public function update(Request $request)
     {
-        $this->validate(
-            request(),
-            [
+        if (auth()->check()) {
+            $this->validate(
+                request(),
+                [
 
-                'skill' => 'required',
+                    'skill' => 'required',
 
-            ]
-        );
+                ]
+            );
 
-        $skill = Skills::find(request('id'));
-        $skill->user_id = auth()->user()->id;
-        $skill->desc = request('skill');
-        $skill->save();
+            $skill = Skills::find(request('id'));
+            $skill->user_id = auth()->user()->id;
+            $skill->desc = request('skill');
+            $skill->save();
 
-        $profiles = Profile::find(auth()->user()->id);
-        $education = Education::all();
-        $skills = Skills::all();
-        return redirect()->to('/employee/education')->with(['profile' => $profiles, 'education' => $education, 'skills' => $skills]);
+            $profiles = Profile::find(auth()->user()->id);
+            $education = Education::all();
+            $skills = Skills::all();
+            return redirect()->to('/employee/education')->with(['profile' => $profiles, 'education' => $education, 'skills' => $skills]);
+        } else {
+            return redirect()->to('/login');
+        }
     }
 
     public function del()
     {
-        $education = Skills::find(request('skill_id'));
-        $education->delete();
-        return redirect()->to('/employee/education')->with('success', 'successfully removed');
+        if (auth()->check()) {
+            $education = Skills::find(request('skill_id'));
+            $education->delete();
+            return redirect()->to('/employee/education')->with('success', 'successfully removed');
+        } else {
+            return redirect()->to('/login');
+        }
     }
 }
